+++ src/routers/v1/EndPollRouter.js
@@ -1,7 +1,7 @@
 // @flow
 import { Request } from 'express';
 import AppDevRouter from '../../utils/AppDevRouter';
-import constants from '../../utils/constants';
+import constants from '../../utils/Constants';
 import SessionsRepo from '../../repos/SessionsRepo';
 
 class EndPollRouter extends AppDevRouter<Object> {
