+++ src/routers/v1/Polls/GetLivePollsRouter.js
@@ -1,7 +1,7 @@
 // @flow
 import { Request } from 'express';
 import AppDevRouter from '../../../utils/AppDevRouter';
-import constants from '../../../utils/constants';
+import constants from '../../../utils/Constants';
 
 class GetLivePollsRouter extends AppDevRouter<Object> {
     constructor() {
