+++ src/routers/v2/Drafts/GetDraftsRouter.js
@@ -1,7 +1,7 @@
 // @flow
 import AppDevEdgeRouter from '../../../utils/AppDevEdgeRouter';
 import DraftsRepo from '../../../repos/DraftsRepo';
-import constants from '../../../utils/constants';
+import constants from '../../../utils/Constants';
 import type { APIDraft } from '../APITypes';
 
 class GetDraftsRouter extends AppDevEdgeRouter<APIDraft> {
