+++ src/routers/v2/Questions/UpdateQuestionRouter.js
@@ -2,7 +2,7 @@
 import { Request } from 'express';
 import AppDevRouter from '../../../utils/AppDevRouter';
 import QuestionsRepo from '../../../repos/QuestionsRepo';
-import constants from '../../../utils/constants';
+import constants from '../../../utils/Constants';
 
 import type { APIQuestion } from '../APITypes';
