+++ src/routers/v2/Users/GetMeRouter.js
@@ -1,7 +1,7 @@
 // @flow
 import { Request } from 'express';
 import AppDevRouter from '../../../utils/AppDevRouter';
-import constants from '../../../utils/constants';
+import constants from '../../../utils/Constants';
 import type { APIUser } from '../APITypes';
 
 class GetMeRouter extends AppDevRouter<APIUser> {
