+++ src/utils/AppDevRouter.js
@@ -8,11 +8,11 @@ import {
     Response,
     NextFunction,
 } from 'express';
-import type { RequestType } from './constants';
+import type { RequestType } from './Constants';
 import AppDevResponse from './AppDevResponse';
 
-import constants from './constants';
-import lib from './lib';
+import constants from './Constants';
+import lib from './Lib';
 
 /**
  * T is the response type for AppDevRouter
